OT97-78/feature/extend-http-users-service - rename users services
